Refactor package decoding.
